requirements.txt
Flask==2.2.5
pyserial==3.5
requests==2.32.3